refactor: LB-前端优化-组件修复&交互优化
--story=116793123
# Reviewed, transaction id: 4968
