+++ front/src/store/loadbalancer.ts
@@ -4,29 +4,33 @@ import { reactive, ref } from 'vue';
 import { useResourceStore } from './resource';
 
 export const useLoadBalancerStore = defineStore('load-balancer', () => {
-  const targetGroupListPageQuery = reactive<IPageQuery>({
-    start: 0,
-    limit: 50,
-  });
   const resourceStore = useResourceStore();
 
   // state - 目标组id
   const targetGroupId = ref('');
-  const allTargetGroupList = ref([]);
-  // state - lb-tree - 当前选中的资源
-  const currentSelectedTreeNode = ref();
-  // state - 目标组id
-
-  // action - lb-tree - 设置当前选中的资源
-  const setCurrentSelectedTreeNode = (node: any) => {
-    // 其中, node 可能为 lb, listener, domain 节点
-    currentSelectedTreeNode.value = node;
-  };
-  // action - 设置目标组id
   const setTargetGroupId = (v: string) => {
     targetGroupId.value = v;
   };
 
+  // state - lb-tree - 当前选中的资源
+  const currentSelectedTreeNode = ref();
+  const setCurrentSelectedTreeNode = (node: any) => {
+    // 其中, node 可能为 lb, listener, domain 节点
+    currentSelectedTreeNode.value = node;
+  };
+
+  // state - 目标组操作场景
+  const currentScene = ref('');
+  const setCurrentScene = (v: string) => {
+    currentScene.value = v;
+  };
+
+  // state - 目标组左侧列表
+  const allTargetGroupList = ref([]);
+  const targetGroupListPageQuery = reactive<IPageQuery>({
+    start: 0,
+    limit: 50,
+  });
   const getTargetGroupList = async () => {
     const [detailRes, countRes] = await Promise.all(
       [false, true].map((isCount) =>
@@ -57,5 +61,7 @@ export const useLoadBalancerStore = defineStore('load-balancer', () => {
     setCurrentSelectedTreeNode,
     getTargetGroupList,
     allTargetGroupList,
+    currentScene,
+    setCurrentScene,
   };
 });
