+++ front/src/views/business/load-balancer/group-view/all-groups-manager/RsConfigTable/index.scss
