+++ front/src/views/business/load-balancer/group-view/all-groups-manager/RsConfigTable/index.tsx
@@ -1,45 +1,47 @@
 import { defineComponent, ref } from 'vue';
-import { SearchSelect, Loading, Table } from 'bkui-vue';
-import useColumns from '@/views/resource/resource-manage/hooks/use-columns';
+// import components
+import { SearchSelect, Loading, Table, Input } from 'bkui-vue';
 import Empty from '@/components/empty';
-import BatchUpdatePopconfirm from '@/components/batch-update-popconfirm';
+import BatchUpdatePopConfirm from '@/components/batch-update-popconfirm';
+// import hooks
+import useColumns from '@/views/resource/resource-manage/hooks/use-columns';
 import './index.scss';
 
 export default defineComponent({
   name: 'RsConfigTable',
   props: {
+    rsList: Array<any>,
     noOperation: Boolean,
     noSearch: Boolean,
-    details: {
-      type: Array,
-      required: true,
-    },
   },
-  emits: ['showAddRsDialog'],
+  emits: ['update:modelValue', 'showAddRsDialog'],
   setup(props, { emit }) {
     // rs配置表单项
     const isTableLoading = ref(false);
     const { columns, settings } = useColumns('rsConfig');
+
     const handleBatchUpdatePort = (_port: number) => {};
     const handleBatchUpdateWeight = (_weight: number) => {};
     const handleDeleteRs = () => {};
+
     const rsTableColumns = [
       ...columns,
       {
         label: () => (
           <>
             <span>端口</span>
-            <BatchUpdatePopconfirm title='端口' onUpdateValue={handleBatchUpdatePort}></BatchUpdatePopconfirm>
+            <BatchUpdatePopConfirm title='端口' onUpdateValue={handleBatchUpdatePort} />
           </>
         ),
         field: 'port',
         isDefaultShow: true,
+        render: () => <Input />,
       },
       {
         label: () => (
           <>
             <span>权重</span>
-            <BatchUpdatePopconfirm title='权重' onUpdateValue={handleBatchUpdateWeight}></BatchUpdatePopconfirm>
+            <BatchUpdatePopConfirm title='权重' onUpdateValue={handleBatchUpdateWeight} />
           </>
         ),
         field: 'weight',
@@ -54,6 +56,7 @@ export default defineComponent({
     // 补充 port 和 weight 的 settings 配置
     settings.value.checked.push('port', 'weight');
     settings.value.fields.push({ label: '端口', field: 'port' }, { label: '权重', field: 'weight' });
+
     return () => (
       <div class='rs-config-table'>
         <div class={`rs-config-operation-wrap${props.noOperation ? ' jc-right' : ''}`}>
@@ -70,7 +73,7 @@ export default defineComponent({
           )}
         </div>
         <Loading loading={isTableLoading.value}>
-          <Table data={props.details} columns={rsTableColumns} settings={settings.value} showOverflowTooltip>
+          <Table data={props.rsList} columns={rsTableColumns} settings={settings.value} showOverflowTooltip>
             {{
               empty: () => {
                 if (isTableLoading.value) return null;
