--- front/src/views/business/load-balancer/group-view/all-groups-manager/target-group-sideslider-content/index.scss
@@ -1,2 +1,0 @@
-.target-group-sideslider-content {
-}
