+++ front/src/views/business/load-balancer/group-view/all-groups-manager/useRenderTGList.tsx
@@ -1,0 +1,137 @@
+// import components
+import { Button, Message } from 'bkui-vue';
+import Confirm from '@/components/confirm';
+// import stores
+import { useAccountStore, useBusinessStore } from '@/store';
+// import custom hooks
+import useColumns from '@/views/resource/resource-manage/hooks/use-columns';
+import useSelection from '@/views/resource/resource-manage/hooks/use-selection';
+import { useTable } from '@/hooks/useTable/useTable';
+// import types
+import { ISearchItem } from 'bkui-vue/lib/search-select/utils';
+// import utils
+import bus from '@/common/bus';
+
+/**
+ * 渲染目标组list
+ */
+export default () => {
+  // use hooks
+  const { columns, settings } = useColumns('targetGroup');
+  const { selections, handleSelectionChange } = useSelection();
+  // use stores
+  const businessStore = useBusinessStore();
+  const accountStore = useAccountStore();
+
+  const searchData: ISearchItem[] = [
+    {
+      id: 'target_group_name',
+      name: '目标组名称',
+    },
+    {
+      id: 'clb_id',
+      name: 'CLB ID',
+    },
+    {
+      id: 'listener_id',
+      name: '监听器ID',
+    },
+    {
+      id: 'vip_address',
+      name: 'VIP地址',
+    },
+    {
+      id: 'vip_domain',
+      name: 'VIP域名',
+    },
+    {
+      id: 'port',
+      name: '端口',
+    },
+    {
+      id: 'protocol',
+      name: '协议',
+    },
+    {
+      id: 'rs_ip',
+      name: 'RS的IP',
+    },
+  ];
+  const tableColumns = [
+    ...columns,
+    {
+      label: '操作',
+      width: 120,
+      render: ({ data }: any) => (
+        <div>
+          <Button text theme={'primary'} onClick={() => handleEditTargetGroup(data)}>
+            编辑
+          </Button>
+          <span
+            v-bk-tooltips={{
+              content: '已绑定了监听器的目标组不可删除',
+              disabled: data.listener_num === 0,
+            }}>
+            <Button
+              text
+              theme={'primary'}
+              disabled={data.listener_num > 0}
+              class={'ml16'}
+              onClick={() => {
+                handleDeleteTargetGroup(data.id, data.name);
+              }}>
+              删除
+            </Button>
+          </span>
+        </div>
+      ),
+    },
+  ];
+
+  const { CommonTable, getListData } = useTable({
+    searchOptions: {
+      searchData,
+    },
+    tableOptions: {
+      columns: tableColumns,
+      extra: {
+        settings: settings.value,
+        onSelect: (selections: any) => {
+          handleSelectionChange(selections, () => true, false);
+        },
+        onSelectAll: (selections: any) => {
+          handleSelectionChange(selections, () => true, true);
+        },
+      },
+    },
+    requestOption: {
+      type: 'target_groups',
+      sortOption: { sort: 'created_at', order: 'DESC' },
+    },
+  });
+
+  // 编辑单个目标组
+  const handleEditTargetGroup = (data: any) => {
+    bus.$emit('editTargetGroup', data);
+  };
+
+  // 删除单个目标组
+  const handleDeleteTargetGroup = (id: string, name: string) => {
+    Confirm('请确定删除目标组', `将删除目标组【${name}】`, () => {
+      businessStore
+        .deleteTargetGroups({
+          bk_biz_id: accountStore.bizs,
+          ids: [id],
+        })
+        .then(() => {
+          getListData();
+          Message({
+            message: '删除成功',
+            theme: 'success',
+          });
+        });
+    });
+  };
+
+  return { searchData, selections, CommonTable, getListData };
+};
