+++ front/src/views/business/load-balancer/group-view/specific-target-group-manager/target-group-detail/index.tsx
@@ -1,7 +1,8 @@
 import { computed, defineComponent } from 'vue';
+// import components
 import { Button } from 'bkui-vue';
+import RsConfigTable from '../../all-groups-manager/RsConfigTable';
 import './index.scss';
-import RsConfigTable from '../../all-groups-manager/rs-config-table';
 
 export default defineComponent({
   name: 'TargetGroupDetail',
